cli/pyproject.toml
[project]
name = "mdcli"
version = "0.1.0"
description = "CLI for indexing and querying markdown files with LLMs"
requires-python = ">=3.10"
dependencies = [
    "typer==0.9.0",
    "httpx==0.26.0",
    "rich==13.7.0",
    "pyyaml==6.0.1",
]

[project.scripts]
mdcli = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
